[test-infra][e2e] make the e2e test replay tool a standalone diem-tool
